Rename file to reflect new umbrella store abstraction
